Move error check to function
